Gradient of the baseline hazard function

Gradient of the baseline hazard function

Usage

basehaz.grad(t, y, parms)

Arguments

t

A non-negative real number specifying the time point at which to evaluate the baseline hazard function.

y

the initial (state) values of the ODE system, which are not used.

parms

A list of parameters containing the following components:

alpha

The coefficients of the Bernstein polynomial.

degree

A non-negative integer specifying the degree of the Bernstein polynomial.

boundary

Boundary knots at which to anchor the Bernstein polynomial

Value

The gradient with respect to the coefficients of the baseline hazard function evaluated at the specified time point.