Hessian Matrix of the Baseline Hazard Function

Hessian Matrix of the Baseline Hazard Function

Usage

basehaz_hess(t, y, parms)

Arguments

t

A non-negative real number specifying the time point at which to evaluate the Hessian matrix of the baseline hazard function.

y

The initial (state) values of the ODE system, which are not used in Cox proportional hazards models.

parms

A list of parameters containing the following components:

alpha

The coefficients of the bernstein polynomial.

n_basis

A positive odd integer specifying the number of bernstein basis functions.

boundary

A numeric vector of length 2 that defines the lower and upper bounds for normalization.

Value

The Hessian matrix with respect to the coefficients of the baseline hazard function evaluated at the specified time point.