Cross-validation for coxstream objects

Cross-validation for coxstream objects

Usage

cv.coxstream(
  formula,
  data,
  boundary,
  idx_col,
  nu = 0.2,
  n_folds = 10,
  n_alphas = 5,
  n_basis_min = 3,
  seed = 0
)

Arguments

formula

A formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

A data frame containing the variables in the model.

boundary

A vector of length 2 containing the boundary knots.

idx_col

The column name of the index column, which is used to distinguish different patients.

nu

A numeric value used to calculate the number of basis functions. Default is 0.2.

n_folds

A positive integer specifying the number of folds for cross-validation. Default is 10.

n_alphas

A positive integer specifying the number of alpha values to test. Default is 5

n_basis_min

A positive integer specifying the minimum number of basis functions. Default is 3.

seed

An integer specifying the random seed for reproducibility. Default is 0.

Value

An object of class cv.coxstream is returned.