Streaming Cox model with repeated observations

Usage

coxstream(formula, data, n_basis, boundary, subject_col, scale = 2, ...)

Arguments

formula

A formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

A data frame containing the variables in the model.

n_basis

A positive integer specifying the number of basis functions.

boundary

A vector of length 2 containing the boundary knots.

subject_col

The column name of the index column, which is used to distinguish different patients.

scale

The scaling factor for the pre-estimated coefficients.

...

Additional arguments (not used).

Value

Additional arguments passed to the optimization function.

An object of class coxstream containing the fitted model. The object contains the following components:

logLik

The log-likelihood of the fitted model.

theta

The estimated coefficients.

hessian

The Hessian matrix of the fitted model.

n_basis

The number of basis functions used.

n_features

The number of features in the model.

n_samples

The number of samples in the model.

formula

The formula used to fit the model.

boundary

The boundary knots used in the model.

subject_col

The index column used to distinguish different patients.

scale

The scaling factor used in the model.

time_stored

The survival times stored for each patient.

call

The matched call.

Details

Fits a streaming Cox proportional hazards model that accommodates repeated observations for subjects.

This function employs the nlm optimization method to estimate the model parameters. If the optimization fails to converge, an error is raised. During the pre-estimation stage, parameters are projected to mitigate bias introduced in the early stage.

See also

Examples

formula <- Surv(time, status) ~ X1 + X2 + X3 + X4 + X5
fit <- coxstream(
  formula, sim[sim$batch_id == 1, ],
  n_basis = 5, boundary = c(0, 3), subject_col = "patient_id"
)
for (batch in 2:10) {
  fit <- update(fit, sim[sim$batch_id == batch, ])
}
summary(fit)
#> Call:
#> update.coxstream(object = fit, newdata = sim[sim$batch_id == 
#>     batch, ])
#> 
#> Number of basis functions:  5 
#> 
#>       coef exp(coef)      se     z      p    
#> X1 0.99023   2.69186 0.05042 19.64 <2e-16 ***
#> X2 1.02152   2.77742 0.05334 19.15 <2e-16 ***
#> X3 1.01102   2.74841 0.05337 18.94 <2e-16 ***
#> X4 0.89143   2.43861 0.05173 17.23 <2e-16 ***
#> X5 1.03534   2.81607 0.05054 20.48 <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>    exp(coef) exp(-coef) lower .95 upper .95
#> X1 2.6919    0.3715     2.4386    2.9714   
#> X2 2.7774    0.3600     2.5017    3.0835   
#> X3 2.7484    0.3638     2.4754    3.0515   
#> X4 2.4386    0.4101     2.2035    2.6988   
#> X5 2.8161    0.3551     2.5505    3.1093