Streaming Cox model with repeated observations

Usage

coxstream(
  formula,
  data,
  n_basis,
  boundary,
  idx_col,
  scale = 2,
  alpha = 2,
  nu = 0.2,
  ...
)

Arguments

formula

A formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

A data frame containing the variables in the model.

n_basis

A positive integer specifying the number of basis functions.

boundary

A vector of length 2 containing the boundary knots.

idx_col

The column name of the index column, which is used to distinguish different patients.

scale

The scaling factor for the pre-estimated coefficients.

alpha

A numeric value used to calculate the number of basis functions. Default is 2.0.

nu

A numeric value used to calculate the number of basis functions. Default is 0.2.

...

Additional arguments (not used).

Value

An object of class coxstream is returned.

Examples

library(coxstream)
formula <- Surv(time, status) ~ X1 + X2 + X3 + X4 + X5
fit <- coxstream(
  formula, sim[sim$batch_id == 1, ],
  n_basis = 3, boundary = c(0, 3), idx_col = "patient_id"
)
for (batch in 2:10) {
  fit <- update(fit, sim[sim$batch_id == batch, ])
}
summary(fit)
#> Call:
#> coxstream(formula = formula, data = sim[sim$batch_id == 1, ], 
#>     n_basis = 3, boundary = c(0, 3), idx_col = "patient_id")
#> 
#> Number of basis:  7 
#>       coef exp(coef)      se     z      p    
#> X1 0.98898   2.68848 0.05123 19.30 <2e-16 ***
#> X2 1.01042   2.74675 0.05364 18.84 <2e-16 ***
#> X3 1.01112   2.74869 0.05431 18.62 <2e-16 ***
#> X4 0.89296   2.44234 0.05207 17.15 <2e-16 ***
#> X5 1.03635   2.81891 0.05106 20.30 <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>    exp(coef) exp(-coef) lower .95 upper .95
#> X1 2.6885    0.3720     2.4317    2.9724   
#> X2 2.7467    0.3641     2.4726    3.0512   
#> X3 2.7487    0.3638     2.4711    3.0574   
#> X4 2.4423    0.4094     2.2054    2.7048   
#> X5 2.8189    0.3547     2.5505    3.1156