Streaming Cox model with repeated observations

Usage

coxstream(formula, data, degree, boundary, idx_col)

Arguments

formula

A formula expression as for regression models, of the form response ~ predictors. The response must be a survival object as returned by the Surv function.

data

A data frame containing the variables in the model.

degree

The degree of the Bernstein polynomial.

boundary

A vector of length 2 containing the boundary knots.

idx_col

The column name of the index column, which is used to distinguish different patients.

Value

An object of class coxstream.

Examples

library(coxstream)
formula <- survival::Surv(time, status) ~ X1 + X2 + X3 + X4 + X5
fit <- coxstream(
  formula, sim[sim$batch_id == 1, ],
  degree = 3, boundary = c(0, 3), idx_col = "patient_id"
)
for (batch in 2:10) {
  fit <- update(fit, sim[sim$batch_id == batch, ])
}
summary(fit)
#> Call:
#> coxstream(formula = formula, data = sim[sim$batch_id == 1, ], 
#>     degree = 3, boundary = c(0, 3), idx_col = "patient_id")
#> 
#>       coef exp(coef)      se     z      p    
#> X1 1.00875   2.74217 0.05174 19.50 <2e-16 ***
#> X2 1.03520   2.81568 0.05387 19.21 <2e-16 ***
#> X3 1.02583   2.78941 0.05478 18.73 <2e-16 ***
#> X4 0.92393   2.51916 0.05214 17.72 <2e-16 ***
#> X5 1.04544   2.84464 0.05139 20.34 <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>    exp(coef) exp(-coef) lower .95 upper .95
#> X1 2.7422    0.3647     2.4777    3.0348   
#> X2 2.8157    0.3552     2.5335    3.1293   
#> X3 2.7894    0.3585     2.5054    3.1056   
#> X4 2.5192    0.3970     2.2744    2.7902   
#> X5 2.8446    0.3515     2.5721    3.1461